Move daemon mock into separate file and package
